+++ ClssLib/Hoa_Don_Chi_Tiet.cs
@@ -12,9 +12,9 @@ namespace ClssLib
     {
         public Guid ID { get; set; }
         public string ma { get; set; }
-        public decimal gia { get; set; }
+        public double gia { get; set; }
         public int trang_thai { get; set; }
-        public decimal thanh_tien { get; set; }
+        public double thanh_tien { get; set; }
         public int so_luong { get; set; }
         public DateTime ngay_tao { get; set; }
         public DateTime? ngay_sua { get; set; }
