+++ WebBanGiayOnline/Models/ViewModel/CartItem.cs
@@ -5,9 +5,9 @@
         public Guid id { get; set; }
         public int Mahh { get; set; }
         public string TenHH { get; set; }
-        public decimal DonGia { get; set; }
+        public double DonGia { get; set; }
         public int SoLuong { get; set; }
 
-        public decimal ThanhTien => SoLuong * DonGia;
+        public double ThanhTien => SoLuong * DonGia;
     }
 }
